This script was written on-site for a specific clutser layout of 8 SU
And then hacked up for a different config being tested.

So it doesn't really work, but it illustrates a number of interesting
points for clusters with Asetek cooling and Vertiv/Emerson power.  Using
SNMP to gather stats from both sets of rack level devices

This also depends on having /etc/snmp/snmp.conf defaults configured
so you don't have to specify a community (public) or SNMP version (2c)
on the snmpget command lines.  You also have to have the EES-POWER and
ASETEK MIB's installed or refrenced.